+++ subscriptions-ws/services/movieService.js
@@ -1,19 +1,35 @@
 // services/movieService.js
-const movieRepository = require('../repositories/movieRepository');
+const { validateMovie } = require("../validations/movieValidation");
+const movieRepository = require("../repositories/movieRepository");
+const AppError = require("../exceptions/AppError");
 
 const getAllMovies = async () => {
   return await movieRepository.getAllMovies();
 };
 
 const createMovie = async (movieData) => {
+  const errors = validateMovie(movieData);
+  if (errors.length) throw new AppError("Validation error", 400, errors);
+  const existingMovie = await movieRepository.getMovieByName(movieData.name);
+  if (existingMovie)
+    throw new AppError(`Movie with name ${movieData.name} already exists`, 400);
+
   return await movieRepository.createMovie(movieData);
 };
 
 const updateMovie = async (id, updateData) => {
+  const errors = validateMovie(updateData);
+  if (errors.length) throw new AppError("Validation error", 400, errors);
+  const existingMovie = await movieRepository.getMovieById(id);
+  if (!existingMovie) throw new AppError(`Movie with ID ${id} not found`, 404);
+
   return await movieRepository.updateMovie(id, updateData);
 };
 
 const deleteMovie = async (id) => {
+  const existingMovie = await movieRepository.getMovieById(id);
+  if (!existingMovie) throw new AppError(`Movie with ID ${id} not found`, 404);
+
   return await movieRepository.deleteMovie(id);
 };
 
@@ -21,5 +37,5 @@ module.exports = {
   getAllMovies,
   createMovie,
   updateMovie,
-  deleteMovie
+  deleteMovie,
 };
